==> 0_DataPreparation/6.1_weather_parameters.ipynb <==
# import packages
import pandas as pd

# load data (latest version of output from script/notebook before) - "5_has_wettercode"

df_wetter = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/5_has_wettercode.csv")
df_no_wetter = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/5_missing_wettercode.csv")

print("df_wetter:", "\n", df_wetter.head(), "\n")
print("df_no_wetter:", "\n",df_no_wetter.head(), "\n")

### Merge data to have all together again (valid values and NaN):

# concatenate dataframes
merged_df = pd.concat([df_wetter, df_no_wetter])

# reset index to ensure the index is sequential after concatenation
merged_df.reset_index(drop=True, inplace=True)

merged_df["Datum"] = pd.to_datetime(merged_df["Datum"])

print(merged_df.head(), "\n")
print(merged_df.info())

### Add precipitation data from external source:

DWD: <br>
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/ <br>
Station: Kiel Holtenau

# Define the path to additional weather data from DWD
file_path = "/workspaces/bakery_sales_prediction/sourcedata/produkt_klima_tag_19740101_20231231_02564.txt"

# Read the file into a DataFrame, parse dates, and treat -999 as NaN
df_prec = pd.read_csv(file_path, sep=';', na_values='-999', parse_dates=['MESS_DATUM'])
df_prec["MESS_DATUM"] = pd.to_datetime(df_prec["MESS_DATUM"])
print(f"Niederschlagsdaten: \n{df_prec.head()}")

# Rename the MESS_DATUM column to Datum
df_prec.rename(columns={'MESS_DATUM': 'Datum', ' RSK': 'Niederschlag'}, inplace=True)

# Select relevant columns
df_prec = df_prec[['Datum', 'Niederschlag']]

# Display the first few rows of the DataFrame
print(f"\nNiederschlagsdaten Auswahl: \n{df_prec.head()}")

# Perform an inner merge to keep only the rows with matching dates
merged_df = pd.merge(merged_df, df_prec, on='Datum', how='left')

# Display the merged DataFrame
print(f"\nNeuer df Wetter mit Niederschlag: \n{merged_df}")

# Check for statistics
print(f"\nStatistics for precipitation: \n{merged_df['Niederschlag'].describe()} \n")

print(merged_df.info())


### Ergebnis speichern:

print(merged_df)

# Save the updated data to new CSV files
merged_df.to_csv('../sourcedata/cleaned_data/6.1_weather_parameters.csv', sep=',', index=False)
