==> 0_DataPreparation/6.4_temperature_deviation.ipynb <==
# import packages
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.impute import KNNImputer

# load data (latest version of output from script/notebook before) - "6.3_weather_class"

df = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/6.3_weather_class.csv")

print("df:", "\n", df.head(), "\n")


### (Seasonal) Temperature deviation

We tried to includ a sudden change in temperature as a variable because a sudden weather change might influence the behaviour of people.

# Step 1: Calculate the 7-day rolling average for temperature
#df['temp_7d_avg'] = df['Temperatur'].rolling(window=7).mean()

# Step 1: Calculate the 7-day rolling average excluding the current day
df['temp_7d_avg'] = df['Temperatur'].shift(1).rolling(window=7).mean()

# Step 2: Fill NA values created by rolling average
df['temp_7d_avg'].fillna(method='bfill', inplace=True)

# Step 3: Calculate the difference for the next day
#df['temp_diff_next_day'] = df['Temperatur'].shift(-1) - df['temp_7d_avg']

# Step 3: Calculate the difference for the current day
df['temp_dv'] = df['Temperatur'] - df['temp_7d_avg']

# Step 4: Create 'warmer than week before' and 'colder than week before' columns
df['warmer_than_week_before'] = df['temp_dv'].apply(lambda x: x if x > 0 else 0)
df['colder_than_week_before'] = df['temp_dv'].apply(lambda x: -x if x < 0 else 0)

# Display the first few rows to check the results
print(df[['Temperatur', 'temp_7d_avg', 'temp_dv', 'warmer_than_week_before', 'colder_than_week_before']].head(15))
df.describe()


print("Missing values in each column:")
print(df.isna().sum())

# Save the updated data to new CSV files
df.to_csv('../sourcedata/cleaned_data/6.4_temperature_deviation.csv', sep=',', index=False)
