==> 0_DataPreparation/6.5_date_variables.ipynb <==
# import packages
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.impute import KNNImputer

# load data (latest version of output from script/notebook before) - "6.4_temperature_deviation"

df = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/6.4_temperature_deviation.csv")

print(df.head(), "\n")

### Create variables from date

df_w = df.copy()

# Ensure the 'Datum' column is in datetime format
df_w['Datum'] = pd.to_datetime(df_w['Datum'], errors='coerce')

# Check if 'Datum' column has been converted correctly
print(df_w.dtypes)

# extract date features
df_w['year'] = df_w['Datum'].dt.year
df_w['month'] = df_w['Datum'].dt.month
df_w['week'] = df_w['Datum'].dt.isocalendar().week
df_w['weekday'] = df_w['Datum'].dt.dayofweek
df_w['day_month'] = df_w['Datum'].dt.day
print(df_w.head())


print(df_w["month"].describe())

# define the season
def get_season(month):
	    if month in [12, 1, 2]:
	        return 'winter', 4
	    elif month in [3, 4, 5]:
	        return 'spring', 1
	    elif month in [6, 7, 8]:
	        return 'summer', 2
	    else:
	        return 'autumn', 3
	
df_w['season_str'], df_w['season'] = zip(*df_w['month'].apply(get_season))


# Create binary features for seasons --> creates one column per season and then 0/1
# probably to generic for predictions?

#weather_data['winter'] = (weather_data['season'] == 'winter').astype(int)
#weather_data['spring'] = (weather_data['season'] == 'spring').astype(int)
#weather_data['summer'] = (weather_data['season'] == 'summer').astype(int)
#weather_data['autumn'] = (weather_data['season'] == 'autumn').astype(int)


print(df_w.head())


# dataframe after the previous steps:
df_w


### Save result:

 # Save the updated data to new CSV files
df_w.to_csv('../sourcedata/cleaned_data/6.5_date_variables.csv', sep=',', index=False)
