==> 0_DataPreparation/7_data_transformation.ipynb <==
# import packages
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# load data

df = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/6.5_date_variables.csv")
print(df.head(), "\n")

# Drop the columns that we do not want as model input:
columns_to_drop = ['week', 'season_str', 'temp_7d_avg', 'warmer_than_week_before', 'colder_than_week_before']
cleaned_df = df.drop(columns=columns_to_drop)
print(cleaned_df.head())
cleaned_df

# Save the updated data to new CSV files
cleaned_df.to_csv('../sourcedata/cleaned_data/7_transformed_variables.csv', sep=',', index=False)
