==> 0_DataPreparation/9_preparation_for_model.ipynb <==
# import packages
import numpy as np
import pickle
import os

# load data: is "8_encoded_variables" the latest version until now?

# Open the pickle file in binary mode
with open("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/8_encoded_variables.pkl", 'rb') as f:
    # Load the data from the pickle file
    prepared_data = pickle.load(f)

#data = pd.read_csv("/workspaces/bakery_sales_prediction/sourcedata/cleaned_data/8_encoded_variables.pkl")
#print(data.head(), "\n")

prepared_data.head()

### Selection of Training, Test and Validation data:

# Set a random seed for reproducibility
np.random.seed(42)

# Shuffle the data
prepared_data = prepared_data.sample(frac=1).reset_index(drop=True)

# Calculate the number of rows for each dataset
n_total = len(prepared_data)
n_training = int(0.7 * n_total)
n_validation = int(0.20 * n_total)

# Split the features and labels for training, validation, and test
# (no test data for our task!)
training_data = prepared_data.iloc[:n_training]
validation_data = prepared_data.iloc[n_training:n_training+n_validation]
test_data = prepared_data.iloc[n_training+n_validation:]

# Separating features and labels
training_features = training_data.drop('Umsatz', axis=1)
validation_features = validation_data.drop('Umsatz', axis=1)
test_features = test_data.drop('Umsatz', axis=1)

training_labels = training_data[['Umsatz']]
validation_labels = validation_data[['Umsatz']]
test_labels = test_data[['Umsatz']]

# Ensure 'Datum' is not used as a feature
training_features = training_features.drop('Datum', axis=1)
validation_features = validation_features.drop('Datum', axis=1)
test_features = test_features.drop('Datum', axis=1)

# Print dimensions of the dataframes
print("Training features dimensions:", training_features.shape)
print("Validation features dimensions:", validation_features.shape)
print("Test features dimensions:", test_features.shape)
print()
print("Training labels dimensions:", training_labels.shape)
print("Validation labels dimensions:", validation_labels.shape)
print("Test labels dimensions:", test_labels.shape)


# Data Export:

# Create subdirectory for the pickle files
subdirectory = "/workspaces/bakery_sales_prediction/sourcedata/pickle_data"
os.makedirs(subdirectory, exist_ok=True)

# Export of the prepared data to subdirectory as pickle files
training_features.to_pickle(f"{subdirectory}/training_features.pkl")
validation_features.to_pickle(f"{subdirectory}/validation_features.pkl")
test_features.to_pickle(f"{subdirectory}/test_features.pkl")
training_labels.to_pickle(f"{subdirectory}/training_labels.pkl")
validation_labels.to_pickle(f"{subdirectory}/validation_labels.pkl")
test_labels.to_pickle(f"{subdirectory}/test_labels.pkl")
